+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # CHANGELOG
 
+## v3.0.0 - 2020-07-09
+
+- Thanks to PA Inc for updating the server scripts to support all game modes in Galactic War
+- Implemented support for:
+  - bounty mode (bounties)
+  - land anywhere (big spawns)
+  - sudden death mode (team death)
+- Replaced my shuffle solution with the new native spawn shuffling solution
+  - Wars pre-v3 will not shuffle spawns
+- Made landing zone shuffling optional
+- Fixed gas giants being included in the planetary intelligence surface area count
+- Updated difficulty tooltip recommendations
+- Show the system's name in place of the primary AI's name
+- Systems installed via Shared Systems for Galactic War will now show their description on the galaxy map
+- Added new option "Tougher Commanders" for those who want the enemy to be harder to kill (TITANS only)
+- Fixed bosses not spawning multiple commanders
+
 ## v2.3.3 - 2020-07-04
 
 - Add support for Default personality tag to GW-CUSTOM